# PortalMobileApp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ITC Student Portal API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Get all buildings",
                "operationId": "f29d20a22fd3a8295e27f96776d5007f",
                "responses": {
                    "200": {
                        "description": "List of buildings"
                    }
                }
            },
            "post": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Create a new building",
                "operationId": "c8f8399919e01b1e4e2e2f6a9f1f5107",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Main Building"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "ITC Campus"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "B01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Building created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/buildings/{id}": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Get building by ID",
                "operationId": "7c2e9a5ec44b127b681717d1d5b91152",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Building detail"
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Update building",
                "operationId": "0bd50384a833589efa669b51b8e1ff88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Building"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "New Address"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "B02"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Building updated"
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Delete building",
                "operationId": "db5360c892964cf865373a95f5da4633",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Building deleted"
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            }
        },
        "/api/classes": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Get all classrooms",
                "operationId": "0be89a41ae5c3b49c014421e66d29455",
                "responses": {
                    "200": {
                        "description": "List of classrooms with building"
                    }
                }
            },
            "post": {
                "tags": [
                    "Classes"
                ],
                "summary": "Create a new classroom",
                "operationId": "3476275874732c00644d08ba4c1f4d67",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "building_id",
                                    "room_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Room A101"
                                    },
                                    "building_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "floor": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_type": {
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "lab",
                                            "meeting",
                                            "other"
                                        ],
                                        "example": "lab"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Classroom created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/classes/{id}": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Get classroom by ID",
                "operationId": "1ed85ef10177abaf60d4a2927655471b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Classroom detail with building"
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Classes"
                ],
                "summary": "Update classroom",
                "operationId": "81c91bac2441ee80a15391a7bee1e794",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Room B202"
                                    },
                                    "building_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "floor": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "room_type": {
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "lab",
                                            "meeting",
                                            "other"
                                        ],
                                        "example": "normal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Classroom updated"
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Classes"
                ],
                "summary": "Delete classroom",
                "operationId": "8b13363f09c96616c06c86a5f63c1beb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Classroom deleted"
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}